Alert delivery methods are not configurable in the spec.
